wallet-api: Get rid of .TH modules
