+++ marlowe/src/Language/Marlowe/Common.hs
@@ -98,8 +98,8 @@ import           Ledger                         ( PubKey(..)
                                                 , Signature(..)
                                                 , Slot(..)
                                                 )
-import qualified Ledger.Ada.TH                  as Ada
-import           Ledger.Ada.TH                  (Ada)
+import qualified Ledger.Ada                  as Ada
+import           Ledger.Ada                  (Ada)
 import           Ledger.Interval                (Interval(..))
 import           Ledger.Validation
 import qualified Ledger.Validation              as Validation
