+++ plutus-use-cases/src/Language/PlutusTx/Coordination/Contracts/CrowdFunding.hs
@@ -32,7 +32,7 @@ import qualified Language.PlutusTx.Prelude   as P
 import           Ledger
 import           Ledger.Validation           as V
 import           Ledger.Value                (Value)
-import qualified Ledger.Value.TH             as VTH
+import qualified Ledger.Value                as VTH
 import           Wallet                      as W
 import qualified Wallet.Emulator             as EM
 import           Wallet.Emulator             (Wallet)
