+++ plutus-wallet-api/plutus-wallet-api.cabal
@@ -52,12 +52,9 @@ library
         Data.Aeson.Extras,
         Ledger,
         Ledger.Ada,
-        Ledger.Ada.TH,
         Ledger.AddressMap,
         Ledger.Map,
-        Ledger.Map.TH,
         Ledger.These,
-        Ledger.These.TH,
         Ledger.Blockchain,
         Ledger.Crypto,
         Ledger.Slot,
@@ -68,7 +65,6 @@ library
         Ledger.Index,
         Ledger.Interval,
         Ledger.Value,
-        Ledger.Value.TH,
         LedgerBytes
     default-language: Haskell2010
     build-depends:
@@ -99,12 +95,9 @@ library plutus-ledger
         Data.Aeson.Extras
         Ledger
         Ledger.Ada
-        Ledger.Ada.TH
         Ledger.AddressMap
         Ledger.Map
-        Ledger.Map.TH
         Ledger.These
-        Ledger.These.TH
         Ledger.Blockchain
         Ledger.Crypto
         Ledger.Slot
@@ -115,7 +108,6 @@ library plutus-ledger
         Ledger.Index
         Ledger.Interval
         Ledger.Value
-        Ledger.Value.TH
         LedgerBytes
     build-depends:
         base >=4.9 && <5,
