+++ plutus-wallet-api/test/Spec.hs
@@ -29,8 +29,8 @@ import           Ledger
 import qualified Ledger.Ada                 as Ada
 import qualified Ledger.Index               as Index
 import qualified Ledger.Map
+import           Ledger.Value               (CurrencySymbol, Value (Value))
 import qualified Ledger.Value               as Value
-import           Ledger.Value.TH            (CurrencySymbol, Value (Value))
 import           LedgerBytes
 import           Test.Tasty
 import           Test.Tasty.Hedgehog        (testProperty)
